move jwks handler to well known directory for clarity
