+++ packages/http-api/src/sync-search/sync-search-handler.test.ts
@@ -16,7 +16,7 @@ import {
   generateKeyPair
 } from 'jose'
 import { getEncryptionKeys } from '../crypto/keys'
-import { getJwk } from '../jwks/jwks-handler'
+import { getJwk } from '../.well-known/jwks-handler'
 
 describe('POST /registry/sync/search', async () => {
   let server: Hapi.Server
